requirements.txt
asgiref==3.8.1
attrs==24.3.0
cfgv==3.4.0
distlib==0.3.9
dj-database-url==2.3.0
dj-static==0.0.6
Django==5.1.4
django-filter==24.3
django-log-request-id==2.1.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
filelock==3.16.1
gunicorn==23.0.0
identify==2.6.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
nodeenv==1.9.1
packaging==24.2
phonenumberslite==8.13.52
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
sqlparse==0.5.3
static3==0.7.0
typing_extensions==4.12.2
uritemplate==4.1.1
virtualenv==20.28.1
